refactor: post detail 타입 파일 분리
