+++ src/features/posts/api/use-get-post-detail.ts
@@ -1,4 +1,4 @@
-import type { PostDetail } from '@/entities/posts/model/types';
+import type { PostDetail } from '@/entities/posts/model/post-detail';
 import { apiClient } from '@/shared/api/client';
 import { useQuery } from '@tanstack/react-query';
 import { HAIR_CONSULT_POSTING_QUERY_KEY_PREFIX } from '../constants/query-keys';
